+++ cli/cli.js
@@ -294,6 +294,7 @@ class DjinnJS {
                     data = data.replace('REPLACE_WITH_DEFAULT_TRANSITION', this.sites[i].defaultTransition);
                     data = data.replace("'REPLACE_WITH_PJAX_STATUS'", this.sites[i].disablePjax);
                     data = data.replace("'REPLACE_WITH_PREFETCH_STATUS'", this.sites[i].disablePrefetching);
+                    data = data.replace("'REPLACE_WITH_USE_PERCENTAGE", this.sites[i].usePercentage);
                     fs.writeFile(runtimeFile, data, error => {
                         if (error) {
                             reject(error);
@@ -436,6 +437,7 @@ class DjinnJS {
                     defaultTransition: this.config.defaultTransition,
                     disablePjax: this.config.disablePjax,
                     disablePrefetching: this.config.disablePrefetching,
+                    usePercentage: this.config.usePercentage,
                 };
                 configChecker(site)
                     .then(validSite => {